requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.10.0
keyboard>=0.13.5
pyperclip>=1.8.2
pyyaml>=6.0

ollama>=0.3.0
anthropic>=0.30.0
openai>=1.30.0

customtkinter>=5.2.0
pystray>=0.19.5
Pillow>=10.0.0

plyer>=2.1.0

pyinstaller>=6.0.0

pytest>=8.0.0
